# developer-docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Persona.ai Developer Platform API",
    "version": "1.0.0",
    "description": "REST + AG-UI API for integrating your product with Persona's agent infrastructure.\n\nEvery endpoint here is a server-to-server call — see the Integration Guide for where in your own stack to call each one from, and how the same endpoint serves both Project-level and per-end-user requests via the `x-persona-external-user-id` header."
  },
  "servers": [
    {
      "url": "https://api.persona.hasanraiyan.me",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "projectCredential": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "<keyId>.<secret>",
        "description": "Developer Platform Project credential, minted via POST /api/v1/projects/{projectId}/credentials. Send as `Authorization: Bearer <keyId>.<secret>`. Optionally pair with the `x-persona-external-user-id` header to act on behalf of one of the Project's own external users. This credential is a server-side secret — see the Integration Guide."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/developer/agents": {
      "post": {
        "tags": [
          "Developer"
        ],
        "summary": "Create an Agent (Project-owned or ExternalUser-owned)",
        "description": "Which kind of Agent gets created depends on the credential's asserted identity: a bare Project credential creates a Project-owned Agent; pairing it with x-persona-external-user-id creates an Agent owned by that external user.\n",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "systemPrompt",
                  "providerId"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Agent created"
          },
          "400": {
            "description": "Validation error, or invalid Provider"
          }
        }
      },
      "get": {
        "tags": [
          "Developer"
        ],
        "summary": "Discover Agents",
        "description": "Results are always scoped to your own Domain — never leaked across Projects. For a bare Project credential (ProjectMachineContext/ ProjectAdminContext): every Agent in this Project's own Domain, any owner type — \"Project discovery\". For a credential paired with x-persona-external-user-id (ProjectRuntimeContext): `scope=mine` returns only that external user's own Agents; omitting it returns this Domain's public Agents only (\"Project-public browse\"). systemPrompt/providerId are stripped from any result the caller doesn't own.\n",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "mine"
              ]
            },
            "description": "ProjectRuntimeContext only — \"mine\" restricts to the asserted external user's own Agents."
          }
        ],
        "responses": {
          "200": {
            "description": "List of Agents"
          }
        }
      }
    },
    "/api/v1/developer/agents/{agentId}": {
      "get": {
        "tags": [
          "Developer"
        ],
        "summary": "Get an Agent",
        "description": "Same visibility rules as Persona (public/unlisted visible in-Domain, private only to the owner). The owner sees full details; non-owners never see systemPrompt/providerId.",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent"
          },
          "404": {
            "description": "Agent not found (or not visible to this Domain/Subject)"
          }
        }
      },
      "patch": {
        "tags": [
          "Developer"
        ],
        "summary": "Update an Agent (owner only)",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated Agent"
          },
          "401": {
            "description": "Unauthorized (not the owner)"
          },
          "404": {
            "description": "Agent not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Developer"
        ],
        "summary": "Delete (soft-delete) an Agent (owner only)",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent deleted"
          },
          "401": {
            "description": "Unauthorized (not the owner)"
          },
          "404": {
            "description": "Agent not found"
          }
        }
      }
    },
    "/api/v1/developer/agents/bulk-delete": {
      "post": {
        "tags": [
          "Developer"
        ],
        "summary": "Delete multiple Agents in one call",
        "description": "Best-effort batch delete — one blocked/not-found id doesn't abort the rest. Up to 100 ids per request.\n",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ids"
                ],
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "maxItems": 100
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "{ deleted: string[], failed: [{ id, reason }] }"
          },
          "400": {
            "description": "ids missing, empty, or over 100 entries"
          }
        }
      }
    },
...
  },
...
}